requirements.txt
tencentcloud-sdk-python==3.0.426
requests==2.25.1
pytz==2021.1